Major update - integrating with the new version of JsRender, with 'codeless' markup syntax, (See JsRender commit for details).
Note: Some changes will be made to JsViews in coming updates, for a more
consistent syntax and implementation between data linking expression
(data-getfrom etc.) and the codeless tags of JsRender.

For convenience, a folder "previous-version-of-jsviews" is provided with the
demos and source code for the previous template syntax. This will be removed
in the future...